fix: asset service and review changes
