+++ backend/src/controllers/assets.controller.ts
@@ -1,16 +1,16 @@
-import { MUNICIPALITY_ID, WHITELIST_ASSET_TYPES } from '@/config';
+import { MUNICIPALITY_ID } from '@/config';
 import { getApiBase } from '@/config/api-config';
 import { AddressAddressCategoryEnum, Attachment, Errand, Stakeholder, StakeholderTypeEnum } from '@/data-contracts/case-data/data-contracts';
-import { Asset, Status } from '@/data-contracts/partyassets/data-contracts';
-import { AssetWithService, ServiceDetails } from '@/interfaces/asset.interface';
+import { Asset } from '@/data-contracts/partyassets/data-contracts';
+import { AssetWithService } from '@/interfaces/asset.interface';
 import { AttachmentCategory, CaseDataNamespace, ParkingPermitCaseType, StakeholderRole } from '@/interfaces/casedata.interface';
-import { enumTitles, getRjsfSchema } from '@/services/jsonschema.service';
 import { HttpException } from '@/exceptions/HttpException';
 import { RequestWithUser } from '@/interfaces/auth.interface';
 import { ApiResponse } from '@/interfaces/service';
 import { User } from '@/interfaces/users.interface';
 import authMiddleware from '@/middlewares/auth.middleware';
 import ApiService from '@/services/api.service';
+import { isAllowedAsset, isVisibleStatus, toClientAsset, toServiceDetails, toVisibleAssets } from '@/services/asset.service';
 import { getCitizen } from '@/services/citizen.service';
 import { fileUploadOptions } from '@/utils/files/fileUploadOptions';
 import { getRepresentingPartyId } from '@/utils/getRepresentingPartyId';
@@ -126,87 +126,6 @@ export class AssetsController {
     return { data: { success: true }, message: 'ok' };
   }
 
-  // Whitelist of asset types from env (WHITELIST_ASSET_TYPES). Types not listed
-  // (e.g. LICENSE) are never returned over the network.
-  private isAllowedAsset = (asset: Asset): boolean => {
-    return !!asset?.type && WHITELIST_ASSET_TYPES.has(asset.type);
-  };
-
-  // Statuses tied to ongoing or superseded cases. DRAFT belongs to active
-  // errands and REPLACED has been superseded by a newer asset, so neither is
-  // shown to citizens. ACTIVE/EXPIRED/BLOCKED/TEMPORARY are all returned.
-  private static readonly HIDDEN_STATUSES: ReadonlySet<Status> = new Set([Status.DRAFT, Status.REPLACED]);
-
-  private isVisibleStatus = (asset: Asset): boolean => {
-    return !!asset?.status && !AssetsController.HIDDEN_STATUSES.has(asset.status);
-  };
-
-  // Assets are addressed by their id (the only universal identifier — the
-  // external assetId exists for externally generated assets like parking
-  // permits, but not for e.g. paratransit). Skip any asset without one.
-  private isAddressable = (asset: Asset): boolean => {
-    return !!asset?.id;
-  };
-
-  // Strip sensitive/internal upstream fields before returning to the client,
-  // without mutating the upstream API object. The id is kept as the client
-  // identifier; jsonParameters are exposed as mapped service details instead.
-  private toClientAsset = (asset: Asset): Asset => {
-    const clientAsset = { ...asset };
-    delete clientAsset.partyId;
-    delete clientAsset.jsonParameters;
-    return clientAsset;
-  };
-
-  private toVisibleAssets = (assets: Asset[]): Asset[] => {
-    return assets.filter(this.isAllowedAsset).filter(this.isVisibleStatus).filter(this.isAddressable);
-  };
-
-  // Form-data values may be stored as a single id, plain id arrays, or
-  // {key,value} arrays. Normalize them before resolving enum titles.
-  private normalizeArray = (values: unknown): string[] => {
-    const rawValues = Array.isArray(values) ? values : typeof values === 'string' ? [values] : [];
-
-    return rawValues
-      .map(value => {
-        if (typeof value === 'string') return value;
-        if (value && typeof value === 'object') {
-          const entry = value as { value?: unknown; key?: unknown };
-          return typeof entry.value === 'string' ? entry.value : typeof entry.key === 'string' ? entry.key : undefined;
-        }
-        return undefined;
-      })
-      .filter((value): value is string => Boolean(value));
-  };
-
-  // Resolve the human-readable service detail for assets that carry form data
-  // (e.g. paratransit) in their first json parameter. The stored values are
-  // enum ids; titles come from the asset's RJSF schema. Returns undefined when
-  // the asset has no such data.
-  private toServiceDetails = async (asset: Asset, user: User): Promise<ServiceDetails | undefined> => {
-    const param = asset.jsonParameters?.[0];
-    if (!param?.value) return undefined;
-
-    let formData: Record<string, unknown> | null;
-    try {
-      formData = typeof param.value === 'string' ? JSON.parse(param.value) : (param.value as Record<string, unknown>);
-    } catch {
-      return undefined;
-    }
-    if (!formData) return undefined;
-
-    const schema = param.schemaId ? await getRjsfSchema(param.schemaId, user) : null;
-
-    return {
-      restyp: enumTitles(schema, 'type', this.normalizeArray(formData.type)),
-      transportMode: enumTitles(schema, 'transportMode', this.normalizeArray(formData.transportMode)),
-      aids: enumTitles(schema, 'mobilityAids', this.normalizeArray(formData.mobilityAids)),
-      addon: enumTitles(schema, 'additionalAids', this.normalizeArray(formData.additionalAids)),
-      comment: typeof formData.notes === 'string' ? formData.notes : '',
-      isWinterService: formData.isWinterService === 'ja' || formData.isWinterService === true || formData.validityType === 'vinterfardtjanst',
-    };
-  };
-
   @Get('/assets')
   @OpenAPI({ summary: 'Return a list of assets for current representing entity' })
   @UseBefore(authMiddleware)
@@ -231,10 +150,8 @@ export class AssetsController {
         throw new HttpException(500, 'No data from API');
       }
 
-      const assets = this.toVisibleAssets(res.data);
-      const data = await Promise.all(
-        assets.map(async asset => ({ ...this.toClientAsset(asset), service: await this.toServiceDetails(asset, req.user) })),
-      );
+      const assets = toVisibleAssets(res.data);
+      const data = await Promise.all(assets.map(async asset => ({ ...toClientAsset(asset), service: await toServiceDetails(asset, req.user) })));
 
       return { data, message: 'success' };
     } catch (error) {
@@ -264,8 +181,6 @@ export class AssetsController {
     }
 
     try {
-      // The partyassets search filters by assetId, not the internal id, so fetch
-      // the representing party's assets and match on id here.
       const params = {
         partyId: getRepresentingPartyId(representing),
       };
@@ -278,13 +193,13 @@ export class AssetsController {
 
       const asset = res.data.find(a => a.id === id);
 
-      if (!asset || !this.isAllowedAsset(asset) || !this.isVisibleStatus(asset)) {
+      if (!asset || !isAllowedAsset(asset) || !isVisibleStatus(asset)) {
         throw new HttpException(404, 'Asset not found');
       }
 
-      const service = await this.toServiceDetails(asset, req.user);
+      const service = await toServiceDetails(asset, req.user);
 
-      return { data: { ...this.toClientAsset(asset), service }, message: 'success' };
+      return { data: { ...toClientAsset(asset), service }, message: 'success' };
     } catch (error) {
       console.error(error);
       if (error.status === 404) {
